refactor(ui): Move organizations table to "/organizations" route
Move the organization table from the root route to `/organizations` and
leave `/` as a minimal placeholder that links to the organizations table. 

Update existing links from `/` to `/organizations` and update the smoke
test to use the new route.
