+++ ui/src/routes/admin/index.tsx
@@ -66,7 +66,7 @@ const OverviewContent = () => {
   return (
     <div className='space-y-4'>
       <div className='grid grid-cols-1 gap-4 md:grid-cols-2 lg:grid-cols-4'>
-        <Link to='/'>
+        <Link to='/organizations'>
           <StatisticsCard
             title='Organizations'
             icon={() => <List className='h-4 w-4' />}
